requirements.txt
aiomysql==0.0.21
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
pycodestyle==2.7.0
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==0.9.3
python-dotenv==0.19.0
PyYAML==5.4.1
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
Babel==2.9.1
